+++ src/Pim/Bundle/CatalogBundle/Manager/CompletenessManager.php
@@ -114,7 +114,6 @@ class CompletenessManager
         }
     }
 
-
     /**
      * Returns an array containing all completeness info and missing attributes for a product
      *